+++ src/searchtool_gtk/__init__.py
@@ -7,7 +7,7 @@ gi.require_version('Adw', '1')
 
 from .client import basic_client, dmenu_client
 from .collation import PathCollator, SearchToolCollator, StringCollator
-from .config import build_modes_from_file
+from .config import build_modes_from_config_file
 from .entry_point import entry_point
 from .exceptions import SearchToolError, SearchToolValidationError
 from .gui import SearchToolApp