+++ src/searchtool_gtk/entry_point.py
@@ -1,19 +1,21 @@
 import sys
+import warnings
 
-import xdg.BaseDirectory
-
-from .config import build_modes_from_file
+from .config import build_modes_from_config_file
 from .exceptions import SearchToolValidationError
 from .gui import SearchToolApp
 
 
 def entry_point() -> None:
+    warnings.simplefilter('always')
+
     try:
-        config_items = build_modes_from_file(
-            xdg.BaseDirectory.load_first_config('searchtool.json'),
-        )
+        config_items = build_modes_from_config_file()
     except SearchToolValidationError as err:
-        raise SystemExit(err) from err
+        if err.__cause__:
+            raise SystemExit(f'Error: {err}. {err.__cause__}.') from err
+
+        raise SystemExit(f'Error: {err}.') from err
 
     app = SearchToolApp(config_items)
     app.run(sys.argv)