+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.3.1 - 2026-05-26
+
+* Silence resource warnings on `subprocess.Popen`.
+
 ## 2.3.0 - 2026-05-15
 
 * Add a TOML configuration option and deprecate JSON configurations.